backend/requirements.txt
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.11.0
attrs==25.4.0
bcrypt==4.1.2
boto3==1.34.19
botocore==1.34.162
brevo-python==1.2.0
certifi==2025.11.12
distro==1.9.0
Django==5.2.8
django-cors-headers==4.3.1
django-dramatiq==0.11.6
django-filter==25.2
django-redis==5.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
dramatiq==1.15.0
drf-spectacular==0.27.0
gunicorn==21.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
inflection==0.5.1
jiter==0.12.0
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
numpy==2.3.5
openai==2.8.1
packaging==25.0
prometheus_client==0.23.1
psycopg2-binary==2.9.9
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.3
redis==5.0.1
referencing==0.37.0
rpds-py==0.29.0
s3transfer==0.10.4
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.4
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
